+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.2.3] - 2025-09-20
+
 ### Fixed
 - Double title in some widget types with TYPO3 >= 13.4.6
 
@@ -229,7 +231,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release
 
 
-[Unreleased]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v3.2.2...HEAD
+[Unreleased]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v3.2.3...HEAD
+[3.2.3]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v3.2.2...v3.2.3
 [3.2.2]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v3.2.1...v3.2.2
 [3.2.1]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v3.2.0...v3.2.1
 [3.2.0]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v3.1.2...v3.2.0